Widen the value domain from i32 to i64
Every scalar in the engine was a 32-bit signed integer: a `number` constant in
a program, a cell of an input file, a column of a physical row, an aggregate
result, and an argument or result of an embedded `@call`. That is narrower than
the standard Datalog value domain -- Souffle's `RamSigned` is `int64_t` -- and
narrower than the workloads the engine targets. Program-analysis facts, hashes,
addresses and identifiers minted by hashing do not fit in 32 bits, and a value
that does not fit is not reported: the reader parses each cell with
`parse::<i32>()` and drops the row when it fails, so a 33-bit identifier turns
a relation into a smaller relation.

The domain is now one alias, `parsing::Val = i64`, used by the parser's integer
constants, the physical rows (`Row<N>` and `FatRow`), the reader, the
comparison and arithmetic kernels, the aggregation kernels, the planner's call
constants, and the `@call` ABI. It lives in `parsing` because that is the crate
every other crate in the workspace depends on, and it exists so those five
places cannot drift apart. An integer literal outside the domain is now
reported by name instead of unwrapping a parse error.

The `@call` ABI moves with it and its version string is bumped to
`flowlog-call-abi-v2-i64-bool`, so a module compiled against the old ABI hashes
to a different cache entry and is rebuilt rather than loaded. Exported
functions must now take `i64` arguments and return `i64` or `bool`; an `i32`
export is refused with a message naming the ABI. The symbol type becomes
`unsafe extern "C" fn(*const i64, usize, *mut i64) -> i32`, where the returned
`i32` remains the ABI status and is not a value.

`Min`, the semiring used by the specialised min-aggregation path, widens from
`u32` to `u64` so the value cast into it cannot truncate. That cast is still
sign-lossy, which is a separate defect fixed in its own commit; this one is a
mechanical widening and preserves existing behaviour.

+++ src/executing/tests/call_e2e.rs
@@ -38,19 +38,19 @@ fn embedded_rust_calls_bind_filter_chain_and_reuse_the_native_cache() {
         r#".code rust
 use std::cmp::min;
 
-fn absolute(value: i32) -> i32 {
+fn absolute(value: i64) -> i64 {
     value.saturating_abs()
 }
 
-pub fn normalize(value: i32, limit: i32) -> i32 {
+pub fn normalize(value: i64, limit: i64) -> i64 {
     min(absolute(value), limit)
 }
 
-pub fn add(value: i32, increment: i32) -> i32 {
+pub fn add(value: i64, increment: i64) -> i64 {
     value.saturating_add(increment)
 }
 
-pub fn keep(value: i32) -> bool {
+pub fn keep(value: i64) -> bool {
     value % 2 == 0
 }
 .endcode
@@ -134,11 +134,11 @@ fn embedded_rust_call_participates_in_a_recursive_fixed_point() {
     fs::write(
         &program_path,
         r#".code rust
-pub fn next(value: i32) -> i32 {
+pub fn next(value: i64) -> i64 {
     value.saturating_add(1)
 }
 
-pub fn at_most(value: i32, limit: i32) -> bool {
+pub fn at_most(value: i64, limit: i64) -> bool {
     value <= limit
 }
 .endcode
